requirements.txt
Django==1.10.5
django-filter==1.0.1
argparse==1.2.1
coverage==3.7.1
djangorestframework==3.5.3
pep8==1.5.6
wsgiref==0.1.2
Pillow==4.0.0
MySQL-python==1.2.5
uWSGI==2.0.12
